Consolidate .netcoreapp.cs files because System.Collections.* projects is no longer cross-compiled
